refactor: multi-module 작업 - widget 라이브러리에서 사용할 확장 함수 모듈 추가
