+++ extensions/build.gradle
@@ -2,7 +2,6 @@ plugins {
     id 'com.android.library'
     id 'kotlin-android'
     id 'kotlin-kapt'
-    id 'maven-publish'
 }
 
 android {
@@ -31,20 +30,5 @@ android {
 }
 
 dependencies {
-    implementation(project(":extensions"))
     implementation("androidx.appcompat:appcompat:1.2.0")
-    implementation("androidx.constraintlayout:constraintlayout:2.0.4")
-}
-
-afterEvaluate {
-    publishing {
-        publications {
-            release(MavenPublication) {
-                from components.release
-                groupId = 'com.github.sieunju.widget'
-                artifactId = 'view'
-                version = deps.release
-            }
-        }
-    }
 }
