+++ extensions/src/main/java/hmju/widget/extensions/Extensions.kt
@@ -1,4 +1,4 @@
-package hmju.widget.view
+package hmju.widget.extensions
 
 import android.content.Context
 import android.content.res.Resources
@@ -7,16 +7,15 @@ import android.os.Build
 import android.util.DisplayMetrics
 import android.util.TypedValue
 import android.view.WindowManager
-import com.hmju.view.R
 
 @Suppress("unused", "MemberVisibilityCanBePrivate")
-internal object Extensions {
+object Extensions {
 
     /**
      * Convert Dp to Int
      * ex. 5.dp
      */
-    internal val Int.dp: Int
+    val Int.dp: Int
         get() = TypedValue.applyDimension(
             TypedValue.COMPLEX_UNIT_DIP,
             this.toFloat(),
@@ -27,7 +26,7 @@ internal object Extensions {
      * Convert Sp to Int
      * ex. 5.dp
      */
-    internal val Int.sp: Int
+    val Int.sp: Int
         get() = TypedValue.applyDimension(
             TypedValue.COMPLEX_UNIT_SP,
             this.toFloat(),
@@ -38,7 +37,7 @@ internal object Extensions {
      * Convert Dp to Float
      * ex. 5F.sp
      */
-    internal val Float.dp: Float
+    val Float.dp: Float
         get() = TypedValue.applyDimension(
             TypedValue.COMPLEX_UNIT_DIP,
             this,
@@ -49,16 +48,16 @@ internal object Extensions {
      * Convert Sp to Float
      * ex. 5F.sp
      */
-    internal val Float.sp: Float
+    val Float.sp: Float
         get() = TypedValue.applyDimension(
             TypedValue.COMPLEX_UNIT_SP,
             this,
             Resources.getSystem().displayMetrics
         )
 
-    internal val Float.toSize: Float
+    val Float.toSize: Float
         get() = this / Resources.getSystem().displayMetrics.density
-    internal val Int.toSize: Int
+    val Int.toSize: Int
         get() = this / Resources.getSystem().displayMetrics.density.toInt()
 
     /**
@@ -74,7 +73,7 @@ internal object Extensions {
         return actionBarSize
     }
 
-    internal fun Context.getDeviceWidth(): Int {
+    fun Context.getDeviceWidth(): Int {
         val wm = getSystemService(Context.WINDOW_SERVICE) as WindowManager
         return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
             wm.currentWindowMetrics.bounds.width()
@@ -86,7 +85,7 @@ internal object Extensions {
         }
     }
 
-    internal fun Context.getDeviceHeight(): Int {
+    fun Context.getDeviceHeight(): Int {
         val wm = getSystemService(Context.WINDOW_SERVICE) as WindowManager
         return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
             wm.currentWindowMetrics.bounds.height()
@@ -101,7 +100,7 @@ internal object Extensions {
     /**
      * StatusBar Height
      */
-    internal fun Context.getStatusBarHeight(): Int {
+    fun Context.getStatusBarHeight(): Int {
         val id = resources.getIdentifier("status_bar_height", "dimen", "android")
         return if (id > 0) resources.getDimensionPixelSize(id) else -1
     }
@@ -109,7 +108,7 @@ internal object Extensions {
     /**
      * NavigationBar Height
      */
-    internal fun Context.getNavigationBarHeight(): Int {
+    fun Context.getNavigationBarHeight(): Int {
         val id = resources.getIdentifier("navigation_bar_height", "dimen", "android")
         return if (id > 0) resources.getDimensionPixelSize(id) else -1
     }
@@ -117,7 +116,7 @@ internal object Extensions {
     /**
      * is NavigationBar Height
      */
-    internal fun Context.isNavigationBar(): Boolean {
+    fun Context.isNavigationBar(): Boolean {
         val id = resources.getIdentifier("config_showNavigationBar", "bool", "android")
         return id > 0 && resources.getBoolean(id)
     }
@@ -125,7 +124,7 @@ internal object Extensions {
     /**
      * get Real ContentsHeight
      */
-    internal fun Context.getRealContentsHeight(): Int {
+    fun Context.getRealContentsHeight(): Int {
         return getDeviceHeight()
             .minus(getStatusBarHeight())
             .minus(getNavigationBarHeight())
