+++ settings.gradle
@@ -4,3 +4,4 @@ include ':view'
 include ':gesture'
 include ':recyclerview'
 include ':viewpager'
+include ':extensions'
