+++ view/src/main/java/hmju/widget/view/CustomTextView.kt
@@ -14,7 +14,7 @@ import androidx.annotation.StyleRes
 import androidx.appcompat.widget.AppCompatTextView
 import androidx.core.widget.TextViewCompat
 import com.hmju.view.R
-import hmju.widget.view.Extensions.toSize
+import hmju.widget.extensions.Extensions.toSize
 
 /**
  * Description : Corner, Border, AutoTextSize 비/활성화 상태에 따라서도
